// extension/src/components/errorlens/declares.d.ts
export declare interface IELExtensionConfigType {
// ...
  statusBarMessageType:
    | 'activeCursor'
    | 'activeLine'
    | 'closestProblem'
    | 'closestSeverity';
// ...
  lintFilePaths: Record<string, string[] | 'none'>;
// ...
}

export declare type TELExtensionConfig = Readonly<IELExtensionConfigType>;
// ...
export declare type TELGroupedByLineDiagnostics = Record<
  string,
  vscode.Diagnostic[]
>;

// statusBar/statusBarIcons.ts
export declare type TELStatusBarProblemType = 'error' | 'warning';

export declare interface IELStatusBarIconsInit {
  isEnabled: boolean;
  atZero: TELExtensionConfig['statusBarIconsAtZero'];
  useBackground: TELExtensionConfig['statusBarIconsUseBackground'];
  targetProblems: TELExtensionConfig['statusBarIconsTargetProblems'];
}

// statusBar/statusBarMessage.ts
export declare interface IELStatusBarMessageInit {
  isEnabled: boolean;
  colorsEnabled: boolean;
  messageType: TELExtensionConfig['statusBarMessageType'];
  priority: TELExtensionConfig['statusBarMessagePriority'];
  alignment: TELExtensionConfig['statusBarMessageAlignment'];
}